# passr_api_spec.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Passr API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    },
    {
      "url": "https://api.passr.ca"
    },
    {
      "url": "https://dev.api.passr.ca/"
    }
  ],
  "components": {
    "responses": {
      "DeleteSuccess": {
        "description": "Delete success"
      },
      "NotFound": {
        "description": "Not found"
      }
    },
    "schemas": {
      "UserRead": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The user's id"
          },
          "createdAt": {
            "type": "number",
            "description": "The user's creation date"
          },
          "updatedAt": {
            "type": "number",
            "description": "The user's last update date"
          }
        },
        "required": ["id", "createdAt", "updatedAt"]
      },
...
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "description": "The id of the resource",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/user": {
      "post": {
        "tags": ["User"],
        "summary": "Create a new user",
        "operationId": "createUser",
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/user/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "get": {
        "tags": ["User"],
        "summary": "Get a user",
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete a user",
        "operationId": "deleteUser",
        "responses": {
          "200": {
            "$ref": "#/components/responses/DeleteSuccess"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}